src/third_party/WebKit/LayoutTests/http/tests/preload/warning/request-headers.html

<!doctype html>
<html>
<link rel="preload" as="fetch" crossorigin="anonymous" href="resources/hello.txt"></link>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
promise_test(async (t) => {
  await fetch('resources/hello.txt', {headers: {'foo': 'bar'}});
}, 'Preload match fails because the request headers do not match (see the console log).');
</script>
</body>
</html>
