third_party/blink/web_tests/platform/generic/http/tests/preload/warning/request-headers-expected.txt

CONSOLE WARNING: A preload for 'http://127.0.0.1:8000/preload/warning/resources/hello.txt' is found, but is not used because the request headers do not match.
This is a testharness.js-based test.
PASS Preload match fails because the request headers do not match (see the console log).
Harness: the test ran to completion.
